requirements.txt
Django
djangorestframework
django-environ
django-cors-headers
drf-spectacular
python-dotenv
django-filter
djangorestframework-simplejwt
